+++ MessagingService.EmailMessage.DomainEvents/MessagingService.EmailMessage.DomainEvents.csproj
@@ -5,8 +5,8 @@
   </PropertyGroup>
 
   <ItemGroup>
-	<PackageReference Include="Shared" Version="1.3.3" />
-	<PackageReference Include="Shared.DomainDrivenDesign" Version="1.3.3" />
+	<PackageReference Include="Shared" Version="1.3.7" />
+	<PackageReference Include="Shared.DomainDrivenDesign" Version="1.3.7" />
   </ItemGroup>
 
 </Project>
