+++ MessagingService.EmailMessageAggregate/MessagingService.EmailMessageAggregate.csproj
@@ -5,8 +5,9 @@
   </PropertyGroup>
 
   <ItemGroup>
-	  <PackageReference Include="Shared" Version="1.3.3" />
-	  <PackageReference Include="Shared.EventStore" Version="1.3.3" />
+	  <PackageReference Include="Grpc.Net.Client" Version="2.47.0" />
+	  <PackageReference Include="Shared" Version="1.3.7" />
+	  <PackageReference Include="Shared.EventStore" Version="1.3.7" />
   </ItemGroup>
 
   <ItemGroup>
