+++ MessagingService.IntegrationTests/MessagingService.IntegrationTests.csproj
@@ -6,11 +6,11 @@
   </PropertyGroup>
 
   <ItemGroup>
-    <PackageReference Include="ClientProxyBase" Version="1.3.3" />
+    <PackageReference Include="ClientProxyBase" Version="1.3.7" />
     <PackageReference Include="Ductus.FluentDocker" Version="2.10.48" />
     <PackageReference Include="Microsoft.NET.Test.Sdk" Version="17.2.0" />
-    <PackageReference Include="SecurityService.Client" Version="1.1.0" />
-    <PackageReference Include="Shared.IntegrationTesting" Version="1.3.3" />
+    <PackageReference Include="SecurityService.Client" Version="1.1.1" />
+    <PackageReference Include="Shared.IntegrationTesting" Version="1.3.7" />
     <PackageReference Include="Shouldly" Version="4.0.3" />
     <PackageReference Include="SpecFlow.Tools.MsBuild.Generation" Version="3.9.74" />
     <PackageReference Include="SpecFlow.xUnit" Version="3.9.74" />
