+++ MessagingService.SMSMessage.DomainEvents/MessagingService.SMSMessage.DomainEvents.csproj
@@ -5,7 +5,7 @@
   </PropertyGroup>
 
   <ItemGroup>
-    <PackageReference Include="Shared.DomainDrivenDesign" Version="1.3.3" />
+    <PackageReference Include="Shared.DomainDrivenDesign" Version="1.3.7" />
   </ItemGroup>
 
 </Project>
