+++ MessagingService.SMSMessageAggregate/MessagingService.SMSMessageAggregate.csproj
@@ -5,7 +5,8 @@
   </PropertyGroup>
 
   <ItemGroup>
-    <PackageReference Include="Shared.EventStore" Version="1.3.3" />
+    <PackageReference Include="Grpc.Net.Client" Version="2.47.0" />
+    <PackageReference Include="Shared.EventStore" Version="1.3.7" />
   </ItemGroup>
 
   <ItemGroup>
