+++ MessagingService/Bootstrapper/DomainServiceRegistry.cs
@@ -1,0 +1,13 @@
+﻿namespace MessagingService.Bootstrapper;
+
+using BusinessLogic.Services;
+using Lamar;
+using Microsoft.Extensions.DependencyInjection;
+
+public class DomainServiceRegistry : ServiceRegistry
+{
+    public DomainServiceRegistry()
+    {
+        this.AddSingleton<IMessagingDomainService, MessagingDomainService>();
+    }
+}
