+++ MessagingService/Common/Extensions.cs
@@ -1,0 +1,22 @@
+﻿namespace MessagingService.Common;
+
+using System;
+using System.Net.Http;
+using EventStore.Client;
+using Microsoft.Extensions.DependencyInjection;
+
+public static class Extensions
+{
+    public static IServiceCollection AddInSecureEventStoreClient(
+        this IServiceCollection services,
+        Uri address,
+        Func<HttpMessageHandler>? createHttpMessageHandler = null)
+    {
+        return services.AddEventStoreClient((Action<EventStoreClientSettings>)(options =>
+                                                                               {
+                                                                                   options.ConnectivitySettings.Address = address;
+                                                                                   options.ConnectivitySettings.Insecure = true;
+                                                                                   options.CreateHttpMessageHandler = createHttpMessageHandler;
+                                                                               }));
+    }
+}
