+++ MessagingService/Startup.cs
@@ -94,27 +94,12 @@ namespace MessagingService
             TypeProvider.LoadDomainEventsTypeDynamically();
         }
 
-        internal static void ConfigureEventStoreSettings(EventStoreClientSettings settings = null)
+        public static void ConfigureEventStoreSettings(EventStoreClientSettings settings)
         {
-            if (settings == null)
-            {
-                settings = new EventStoreClientSettings();
-            }
-
-            settings.CreateHttpMessageHandler = () => new SocketsHttpHandler
-                                                      {
-                                                          SslOptions =
-                                                          {
-                                                              RemoteCertificateValidationCallback = (sender,
-                                                                                                     certificate,
-                                                                                                     chain,
-                                                                                                     errors) => true,
-                                                          }
-                                                      };
-            settings.ConnectionName = Startup.Configuration.GetValue<String>("EventStoreSettings:ConnectionName");
             settings.ConnectivitySettings = EventStoreClientConnectivitySettings.Default;
             settings.ConnectivitySettings.Address = new Uri(Startup.Configuration.GetValue<String>("EventStoreSettings:ConnectionString"));
             settings.ConnectivitySettings.Insecure = Startup.Configuration.GetValue<Boolean>("EventStoreSettings:Insecure");
+
 
             settings.DefaultCredentials = new UserCredentials(Startup.Configuration.GetValue<String>("EventStoreSettings:UserName"),
                                                               Startup.Configuration.GetValue<String>("EventStoreSettings:Password"));
@@ -127,11 +112,9 @@ namespace MessagingService
         {
             ConfigurationReader.Initialise(Startup.Configuration);
             
-            Startup.ConfigureEventStoreSettings();
-            
-            services.IncludeRegistry<MiddlewareRegistry>();
             services.IncludeRegistry<MediatorRegistry>();
             services.IncludeRegistry<RepositoryRegistry>();
+            services.IncludeRegistry<MiddlewareRegistry>();
             services.IncludeRegistry<DomainServiceRegistry>();
             services.IncludeRegistry<DomainEventHandlerRegistry>();
             services.IncludeRegistry<MessagingProxyRegistry>();
